refactor: move feature flags to their own folder
