+++ .changeset/late-years-doubt.md
@@ -1,0 +1,6 @@
+---
+"@devtools/feature-flags": minor
+"@devtools/shell": minor
+---
+
+Add feature flags types and hooks and add makeTool fixture for tests
